Cache ICrushable traits in actor.
