+++ OpenRA.Mods.Common/Traits/Crushable.cs
@@ -10,6 +10,7 @@
 #endregion
 
 using OpenRA.Primitives;
+using OpenRA.Traits;
 
 namespace OpenRA.Mods.Common.Traits
 {
